Clipboard support for obtaining file path. Creation of a run.bat for running both scripts at the same time.
